requirements.txt
rich
requests
beautifulsoup4
dnspython
selenium
playwright
jinja2
reportlab
asyncio
webdriver-manager
